+++ CHANGELOG.md
@@ -1,7 +1,16 @@
 ## PCBash Changelog
 
+### 2.0.0
+- Changed "runCommandPromise" to "run"
+- Changed "runCommandInMachinePromise" to "runInMachine"
+- Changed from shelljs to execa
+- Changed from promises to async/await
+- Changed from jasmine to jest
+- Changed linting from eslintrc.json to eslintrc.js
+- Added code coverage
 
+### 1.1.1
+- bumped ecma to 8 and added await spec
 
 ### 1.1.0
-
 - 'putStringInFile' also accepts objects